Move hegel properties into separate _hegel_test.go files
The overall diff against main is easier to read this way: the table
tests show up as deleted files and the properties as new files, rather
than a wholesale in-place rewrite of the existing test files.
